refactor to use _get_async_httpx_client
